Bath Module

HierarchicalEOM.Bath.C — Method

C(bath, tlist)

Calculate the correlation function $C(t)$ for a given bosonic bath and time list.

Note that

\[C(t)=\sum_{u=\textrm{R},\textrm{I}}(\delta_{u, \textrm{R}} + i\delta_{u, \textrm{I}})C^{u}(t)\]

Parameters

bath::BosonBath : The bath object which describes a certain bosonic bath.

tlist::AbstractVector: The specific time.

Returns

clist::Vector{ComplexF64} : a list of the value of correlation function according to the given time list.

HierarchicalEOM.Bath.C — Method

C(bath, tlist)

Calculate the correlation function $C^{\nu=+}(t)$ and $C^{\nu=-}(t)$ for a given fermionic bath and time list. Here, $\nu=+$ represents the absorption process and $\nu=-$ represents the emmision process.

Parameters

bath::FermionBath : The bath object which describes a certain fermionic bath.

tlist::AbstractVector: The specific time.

Returns

cplist::Vector{ComplexF64} : a list of the value of the absorption ($\nu=+$) correlation function according to the given time list.

cmlist::Vector{ComplexF64} : a list of the value of the emission ($\nu=-$) correlation function according to the given time list.

Bosonic Bath

HierarchicalEOM.Bath.BosonBath — Type

struct BosonBath <: AbstractBath

An object which describes the interaction between system and bosonic bath

Fields

bath : the different boson-bath-type objects which describes the interaction between system and bosonic bath

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

dim : the dimension of the coupling operator (should be equal to the system dimension).

Nterm : the number of exponential-expansion term of correlation functions

δ : The approximation discrepancy which is used for adding the terminator to HEOM matrix (see function: addTerminator)

Methods

One can obtain the $k$-th exponent (exponential-expansion term) from bath::BosonBath by calling : bath[k]. HierarchicalEOM.jl also supports the following calls (methods) :

bath[1:k];   # returns a vector which contains the exponents from the `1`-st to the `k`-th term.
bath[1:end]; # returns a vector which contains all the exponential-expansion terms
bath[:];     # returns a vector which contains all the exponential-expansion terms
from b in bath
    # do something
end

HierarchicalEOM.Bath.BosonBath — Method

BosonBath(op, η, γ, δ=0.0; combine=true)

Generate BosonBath object for the case where real part and imaginary part of the correlation function are combined.

\[\begin{aligned} C(\tau) &=\frac{1}{2\pi}\int_{0}^{\infty} d\omega J(\omega)\left[n(\omega)e^{i\omega \tau}+(n(\omega)+1)e^{-i\omega \tau}\right]\\ &=\sum_i \eta_i \exp(-\gamma_i \tau), \end{aligned}\]

where $J(\omega)$ is the spectral density of the bath and $n(\omega)$ represents the Bose-Einstein distribution.

Parameters

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

η::Vector{Ti<:Number} : the coefficients $\eta_i$ in bath correlation function $C(\tau)$.

γ::Vector{Tj<:Number} : the coefficients $\gamma_i$ in bath correlation function $C(\tau)$.

δ::Number : The approximation discrepancy (Default to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator)

combine::Bool : Whether to combine the exponential-expansion terms with the same frequency. Defaults to true.

HierarchicalEOM.Bath.BosonBath — Method

BosonBath(op, η_real, γ_real, η_imag, γ_imag, δ=0.0; combine=true)

Generate BosonBath object for the case where the correlation function splits into real part and imaginary part.

\[\begin{aligned} C(\tau) &=\frac{1}{2\pi}\int_{0}^{\infty} d\omega J(\omega)\left[n(\omega)e^{i\omega \tau}+(n(\omega)+1)e^{-i\omega \tau}\right]\\ &=\sum_i \eta_i \exp(-\gamma_i \tau), \end{aligned}\]

where $J(\omega)$ is the spectral density of the bath and $n(\omega)$ represents the Bose-Einstein distribution.

When $\gamma_i \neq \gamma_i^*$, a closed form for the HEOM can be obtained by further decomposing $C(\tau)$ into its real (R) and imaginary (I) parts as

\[C(\tau)=\sum_{u=\textrm{R},\textrm{I}}(\delta_{u, \textrm{R}} + i\delta_{u, \textrm{I}})C^{u}(\tau)\]

where $\delta$ is the Kronecker delta function and $C^{u}(\tau)=\sum_i \eta_i^u \exp(-\gamma_i^u \tau)$

Parameters

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

η_real::Vector{Ti<:Number} : the coefficients $\eta_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

γ_real::Vector{Tj<:Number} : the coefficients $\gamma_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

η_imag::Vector{Tk<:Number} : the coefficients $\eta_i$ in imaginary part of bath correlation function $C^{u=\textrm{I}}$.

γ_imag::Vector{Tl<:Number} : the coefficients $\gamma_i$ in imaginary part of bath correlation function $C^{u=\textrm{I}}$.

δ::Number : The approximation discrepancy (Default to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator)

combine::Bool : Whether to combine the exponential-expansion terms with the same frequency. Defaults to true.

HierarchicalEOM.Bath.bosonReal — Type

struct bosonReal <: AbstractBosonBath

A bosonic bath for the real part of bath correlation function $C^{u=\textrm{R}}$

Fields

Comm : the super-operator (commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

γ : the coefficients $\gamma_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

Nterm : the number of exponential-expansion term of correlation function

HierarchicalEOM.Bath.bosonReal — Method

bosonReal(op, η_real, γ_real)

Generate bosonic bath for the real part of bath correlation function $C^{u=\textrm{R}}$

Parameters

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

η_real::Vector{Ti<:Number} : the coefficients $\eta_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

γ_real::Vector{Tj<:Number} : the coefficients $\gamma_i$ in real part of bath correlation function $C^{u=\textrm{R}}$.

HierarchicalEOM.Bath.bosonImag — Type

struct bosonImag <: AbstractBosonBath

A bosonic bath for the imaginary part of bath correlation function $C^{u=\textrm{I}}$

Fields

Comm : the super-operator (commutator) for the coupling operator.

anComm : the super-operator (anti-commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ in imaginary part of bath correlation function $C^{u=\textrm{I}}$.

γ : the coefficients $\gamma_i$ in imaginary part of bath correlation function $C^{u=\textrm{I}}$.

Nterm : the number of exponential-expansion term of correlation function

HierarchicalEOM.Bath.bosonImag — Method

bosonImag(op, η_imag, γ_imag)

Generate bosonic bath for the imaginary part of correlation function $C^{u=\textrm{I}}$

Parameters

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

η_imag::Vector{Ti<:Number} : the coefficients $\eta_i$ in imaginary part of bath correlation functions $C^{u=\textrm{I}}$.

γ_imag::Vector{Tj<:Number} : the coefficients $\gamma_i$ in imaginary part of bath correlation functions $C^{u=\textrm{I}}$.

HierarchicalEOM.Bath.bosonRealImag — Type

sturct bosonRealImag <: AbstractBosonBath

A bosonic bath which the real part and imaginary part of the bath correlation function are combined

Fields

Comm : the super-operator (commutator) for the coupling operator.

anComm : the super-operator (anti-commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η_real : the real part of coefficients $\eta_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

η_imag : the imaginary part of coefficients $\eta_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

γ : the coefficients $\gamma_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

Nterm : the number of exponential-expansion term of correlation function

HierarchicalEOM.Bath.bosonRealImag — Method

bosonRealImag(op, η_real, η_imag, γ)

Generate bosonic bath which the real part and imaginary part of the bath correlation function are combined

Parameters

op : The system coupling operator, must be Hermitian and, for fermionic systems, even-parity to be compatible with charge conservation.

η_real::Vector{Ti<:Number} : the real part of coefficients $\eta_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

η_imag::Vector{Tj<:Number} : the imaginary part of coefficients $\eta_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

γ::Vector{Tk<:Number} : the coefficients $\gamma_i$ in bath correlation function $\sum_i \eta_i \exp(-\gamma_i t)$.

Fermionic Bath

HierarchicalEOM.Bath.FermionBath — Type

struct FermionBath <: AbstractBath

An object which describes the interaction between system and fermionic bath

Fields

bath : the different fermion-bath-type objects which describes the interaction

op : The system "emission" operator according to the system-fermionic-bath interaction.

dim : the dimension of the coupling operator (should be equal to the system dimension).

Nterm : the number of exponential-expansion term of correlation functions

δ : The approximation discrepancy which is used for adding the terminator to HEOM matrix (see function: addTerminator)

Methods

One can obtain the $k$-th exponent (exponential-expansion term) from bath::FermionBath by calling : bath[k]. HierarchicalEOM.jl also supports the following calls (methods) :

bath[1:k];   # returns a vector which contains the exponents from the `1`-st to the `k`-th term.
bath[1:end]; # returns a vector which contains all the exponential-expansion terms
bath[:];     # returns a vector which contains all the exponential-expansion terms
from b in bath
    # do something
end

HierarchicalEOM.Bath.FermionBath — Method

FermionBath(op, η_absorb, γ_absorb, η_emit, γ_emit, δ=0.0)

Generate FermionBath object

\[\begin{aligned} C^{\nu=+}(\tau) &=\frac{1}{2\pi}\int_{-\infty}^{\infty} d\omega J(\omega) n(\omega) e^{i\omega \tau}\\ &=\sum_i \eta_i^{\nu=+} \exp(-\gamma_i^{\nu=+} \tau),\\ C^{\nu=-}(\tau) &=\frac{1}{2\pi}\int_{-\infty}^{\infty} d\omega J(\omega) (1-n(\omega)) e^{-i\omega \tau}\\ &=\sum_i \eta_i^{\nu=-} \exp(-\gamma_i^{\nu=-} \tau), \end{aligned}\]

where $\nu=+$ ($\nu=-$) represents absorption (emission) process, $J(\omega)$ is the spectral density of the bath and $n(\omega)$ is the Fermi-Dirac distribution.

Parameters

op : The system annihilation operator according to the system-fermionic-bath interaction.

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ of absorption bath correlation function $C^{\nu=+}(\tau)$.

γ_absorb::Vector{Tj<:Number} : the coefficients $\gamma_i$ of absorption bath correlation function $C^{\nu=+}(\tau)$.

η_emit::Vector{Tk<:Number} : the coefficients $\eta_i$ of emission bath correlation function $C^{\nu=-}(\tau)$.

γ_emit::Vector{Tl<:Number} : the coefficients $\gamma_i$ of emission bath correlation function $C^{\nu=-}(\tau)$.

δ::Number : The approximation discrepancy (Defaults to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator)

HierarchicalEOM.Bath.fermionAbsorb — Type

struct fermionAbsorb <: AbstractFermionBath

An bath object which describes the absorption process of the fermionic system by a correlation function $C^{\nu=+}$

Fields

spre : the super-operator (right side operator multiplication) for the coupling operator.

spost : the super-operator (left side operator multiplication) for the coupling operator.

spreD : the super-operator (right side operator multiplication) for the adjoint of the coupling operator.

spostD : the super-operator (left side operator multiplication) for the adjoint of the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ of absorption bath correlation function $C^{\nu=+}$.

γ : the coefficients $\gamma_i$ of absorption bath correlation function $C^{\nu=+}$.

η_emit : the coefficients $\eta_i$ of emission bath correlation function $C^{\nu=-}$.

Nterm : the number of exponential-expansion term of correlation function

HierarchicalEOM.Bath.fermionAbsorb — Method

fermionAbsorb(op, η_absorb, γ_absorb, η_emit)

Generate fermionic bath which describes the absorption process of the fermionic system by a correlation function $C^{\nu=+}$

# Parameters

op : The system creation operator according to the system-fermionic-bath interaction.

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ of absorption bath correlation function $C^{\nu=+}$.

γ_absorb::Vector{Tj<:Number} : the coefficients $\gamma_i$ of absorption bath correlation function $C^{\nu=+}$.

η_emit::Vector{Tk<:Number} : the coefficients $\eta_i$ of emission bath correlation function $C^{\nu=-}$.

HierarchicalEOM.Bath.fermionEmit — Type

struct fermionEmit <: AbstractFermionBath

An bath object which describes the emission process of the fermionic system by a correlation function $C^{\nu=-}$

Fields

spre : the super-operator (right side operator multiplication) for the coupling operator.

spost : the super-operator (left side operator multiplication) for the coupling operator.

spreD : the super-operator (right side operator multiplication) for the adjoint of the coupling operator.

spostD : the super-operator (left side operator multiplication) for the adjoint of the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ of emission bath correlation function $C^{\nu=-}$.

γ : the coefficients $\gamma_i$ of emission bath correlation function $C^{\nu=-}$.

η_absorb : the coefficients $\eta_i$ of absorption bath correlation function $C^{\nu=+}$.

Nterm : the number of exponential-expansion term of correlation function

HierarchicalEOM.Bath.fermionEmit — Method

fermionEmit(op, η_emit, γ_emit, η_absorb)

Generate fermionic bath which describes the emission process of the fermionic system by a correlation function $C^{\nu=-}$

Parameters

op : The system annihilation operator according to the system-fermionic-bath interaction.

η_emit::Vector{Ti<:Number} : the coefficients $\eta_i$ of emission bath correlation function $C^{\nu=-}$.

γ_emit::Vector{Ti<:Number} : the coefficients $\gamma_i$ of emission bath correlation function $C^{\nu=-}$.

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ of absorption bath correlation function $C^{\nu=+}$.

Exponent

HierarchicalEOM.Bath.Exponent — Type

struct Exponent

An object which describes a single exponential-expansion term (naively, an excitation mode) within the decomposition of the bath correlation functions.

The expansion of a bath correlation function can be expressed as : $C(t) = \sum_i \eta_i \exp(-\gamma_i t)$.

Fields

op : The system coupling operator according to system-bath interaction.

η::Number : the coefficient $\eta_i$ in bath correlation function.

γ::Number : the coefficient $\gamma_i$ in bath correlation function.

types::String : The type-tag of the exponent.

The different types of the Exponent:

"bR" : from real part of bosonic correlation function $C^{u=\textrm{R}}(t)$

"bI" : from imaginary part of bosonic correlation function $C^{u=\textrm{I}}(t)$

"bRI" : from combined (real and imaginary part) bosonic bath correlation function $C(t)$

"fA" : from absorption fermionic correlation function $C^{\nu=+}(t)$

"fE" : from emission fermionic correlation function $C^{\nu=-}(t)$