Correlation Functions

Bosonic Bath Correlation Functions

Heom.CorrelationFunc.Boson_DrudeLorentz_Matsubara — Function

Boson_DrudeLorentz_Matsubara(op, λ, W, T, N)

Constructing Drude-Lorentz bosonic bath with Matsubara expansion

Parameters

op : The system operator according to the system-fermionic-bath interaction.

λ::Real: The coupling strength between the system and the bath.

W::Real: The reorganization energy (band-width) of the bath.

T::Real: The temperature of the bath.

N::Int: Number of exponential terms used to approximate the bath correlation functions.

Returns

bath::BosonBath : a bosonic bath object with describes the interaction between system and bosonic bath

Heom.CorrelationFunc.Boson_DrudeLorentz_Pade — Function

Boson_DrudeLorentz_Pade(op, λ, W, T, N)

Constructing Drude-Lorentz bosonic bath with Padé expansion

A Padé approximant is a sum-over-poles expansion (see here for more details).

The application of the Padé method to spectrum decompoisitions is described in Ref. [1].

[1] J. Chem. Phys. 134, 244106 (2011)

Parameters

op : The system operator according to the system-fermionic-bath interaction.

λ::Real: The coupling strength between the system and the bath.

W::Real: The reorganization energy (band-width) of the bath.

T::Real: The temperature of the bath.

N::Int: Number of exponential terms used to approximate the bath correlation functions.

Returns

bath::BosonBath : a bosonic bath object with describes the interaction between system and bosonic bath

Fermionic Bath Correlation Functions

Heom.CorrelationFunc.Fermion_Lorentz_Matsubara — Function

Fermion_Lorentz_Matsubara(op, λ, μ, W, T, N)

Constructing Lorentzian fermionic bath with Matsubara expansion

Parameters

op : The system "emission" operator according to the system-fermionic-bath interaction.

λ::Real: The coupling strength between the system and the bath.

μ::Real: The chemical potential of the bath.

W::Real: The reorganization energy (band-width) of the bath.

T::Real: The temperature of the bath.

N::Int: Number of exponential terms used to approximate the bath correlation functions.

Returns

bath::FermionBath : a fermionic bath object with describes the interaction between system and fermionic bath

Heom.CorrelationFunc.Fermion_Lorentz_Pade — Function

Fermion_Lorentz_Pade(op, λ, μ, W, T, N)

Constructing Lorentzian fermionic bath with Padé expansion

A Padé approximant is a sum-over-poles expansion (see here for more details).

The application of the Padé method to spectrum decompoisitions is described in Ref. [1].

[1] J. Chem. Phys. 134, 244106 (2011)

Parameters

op : The system "emission" operator according to the system-fermionic-bath interaction.

λ::Real: The coupling strength between the system and the bath.

μ::Real: The chemical potential of the bath.

W::Real: The reorganization energy (band-width) of the bath.

T::Real: The temperature of the bath.

N::Int: Number of exponential terms used to approximate the bath correlation functions.

Returns

bath::FermionBath : a fermionic bath object with describes the interaction between system and fermionic bath