Heom API Module

Heom liouvillian superoperator matrices

Boson Heom matrix

Heom.HeomAPI.M_Boson — Type

mutable struct M_Boson <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for bosonic bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

tier : the tier (cutoff) for the bath

dim : the dimension of system

N : the number of total ADOs

Nb : the number of bosonic ADOs

Nf : the number of fermionic ADOs (should be zero)

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix (restrict to :none for boson)

bath::Vector{BosonBath} : the vector which stores all BosonBath objects

hierarchy::HierarchyDict: the object which contains all dictionaries for boson-bath-ADOs hierarchy.

Heom.HeomAPI.M_Boson — Method

M_Boson(Hsys, tier, Bath; verbose=true)

Generate the boson-type Heom liouvillian superoperator matrix

Parameters

Hsys : The system Hamiltonian

tier::Int : the tier (cutoff) for the bath

Bath::Vector{BosonBath} : objects for different bosonic baths

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

Fermion Heom matrix

Heom.HeomAPI.M_Fermion — Type

mutable struct M_Fermion <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for fermionic bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

tier : the tier (cutoff) for the bath

dim : the dimension of system

N : the number of total ADOs

Nb : the number of bosonic ADOs (should be zero)

Nf : the number of fermionic ADOs

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix

bath::Vector{FermionBath} : the vector which stores all FermionBath objects

hierarchy::HierarchyDict: the object which contains all dictionaries for fermion-bath-ADOs hierarchy.

Heom.HeomAPI.M_Fermion — Type

M_Fermion(Hsys, tier, Bath, parity=:even; verbose=true)

Generate the fermion-type Heom liouvillian superoperator matrix

Parameters

Hsys : The system Hamiltonian

tier::Int : the tier (cutoff) for the bath

Bath::Vector{FermionBath} : objects for different fermionic baths

parity::Symbol : The parity symbol of the density matrix (either :odd or :even). Defaults to :even.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

Boson-Fermion Heom matrix

Heom.HeomAPI.M_Boson_Fermion — Type

mutable struct M_Boson_Fermion <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for mixtured (bosonic and fermionic) bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

tier_b : the tier (cutoff) for bosonic bath

tier_f : the tier (cutoff) for fermionic bath

dim : the dimension of system

N : the number of total ADOs

Nb : the number of bosonic ADOs

Nf : the number of fermionic ADOs

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix

ado2idx_b : the bosonic ADO-to-index dictionary

ado2idx_f : the fermionic ADO-to-index dictionary

bath_b::Vector{BosonBath} : the vector which stores all BosonBath objects

bath_f::Vector{FermionBath} : the vector which stores all FermionBath objects

hierarchy_b::HierarchyDict: the object which contains all dictionaries for boson-bath-ADOs hierarchy.

hierarchy_f::HierarchyDict: the object which contains all dictionaries for fermion-bath-ADOs hierarchy.

Heom.HeomAPI.M_Boson_Fermion — Type

M_Boson_Fermion(Hsys, tier_b, tier_f, Bath_b, Bath_f, parity=:even; verbose=true)

Generate the boson-fermion-type Heom liouvillian superoperator matrix

Parameters

Hsys : The system Hamiltonian

tier_b::Int : the tier (cutoff) for the bosonic bath

tier_f::Int : the tier (cutoff) for the fermionic bath

Bath_b::Vector{BosonBath} : objects for different bosonic baths

Bath_f::Vector{FermionBath} : objects for different fermionic baths

parity::Symbol : The parity symbol of the density matrix (either :odd or :even). Defaults to :even.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

Functions for Heom Matrices

Base.size — Method

size(M::AbstractHEOMMatrix)

Returns the size of the Heom liouvillian superoperator matrix

Heom.HeomAPI.addDissipator! — Function

addDissipator!(M, jumpOP)

Adding dissipator to a given HEOM matrix.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

jumpOP::AbstractVector : The collapse (jump) operators to add. Defaults to empty vector [].

Heom.HeomAPI.addTerminator! — Function

addTerminator!(M, Bath)

Adding terminator to a given HEOM matrix.

The terminator is a liouvillian term representing the contribution to the system-bath dynamics of all exponential-expansion terms beyond Bath.Nterm

The difference between the true correlation function and the sum of the Bath.Nterm-exponential terms is approximately 2 * δ * dirac(t). Here, δ is the approximation discrepancy and dirac(t) denotes the Dirac-delta function.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

Bath::Union{BosonBath, FermionBath} : The bath object which contains the approximation discrepancy δ

Auxiliary Density Operators (ADOs)

Heom.HeomAPI.ADOs — Type

mutable struct ADOs

The Auxiliary Density Operators for Heom model.

Fields

data : the vectorized auxiliary density operators

dim : the dimension of the system

Nb : the number of bosonic states

Nf : the number of fermionic states

Methods

For pure bosonic (or fermionic) type bath ADOs, one can obtain the density matrix for specific index (idx) by calling : ados[idx]. Heom.jl also supports the following calls (methods) :

length(ados);  # returns the total number of `ADOs`
ados[1:idx];   # returns a vector which contains the `ADO` (in matrix form) from index `1` to `idx`
ados[1:end];   # returns a vector which contains all the `ADO` (in matrix form)
ados[:];       # returns a vector which contains all the `ADO` (in matrix form)
from rho in ados  # iteration
    # do something
end

For mixed (bosonic and fermionic) type bath ADOs, one needs two indices (idx_b and idx_f), and thus, ados[idx_b, idx_f]. Note that the first index specifies the bosonic bath index while the other one specifies the fermionic bath. Heom.jl also supports the following calls (methods) :

length(ados);       # returns the total number of `ADOs`
ados[idx_b, 1:end]; # returns a vector which contains all the fermionic `ADO` (in matrix form) where bosonic index is `idx_b`
ados[1:end, idx_f]; # returns a vector which contains all the bosonic `ADO` (in matrix form) where bosonic index is `idx_f`
ados[idx_b, :];     # returns a vector which contains all the fermionic `ADO` (in matrix form) where bosonic index is `idx_b`
ados[:, idx_f];     # returns a vector which contains all the bosonic `ADO` (in matrix form) where bosonic index is `idx_f`

But, currently, we don't support iterate() for mixed bath ADOs.

Heom.HeomAPI.ADOs — Type

ADOs(V, Nb=0, Nf=0)

Gernerate the object of auxiliary density operators for Heom model.

Parameters

V::AbstractVector : the vectorized auxiliary density operators

Nb::Int : the number of bosonic states. Defaults to 0.

Nf::Int : the number of fermionic states Defaults to 0.

Functions for Auxiliary Density Operators

Base.length — Method

length(A::ADOs)

Returns the total number of the Auxiliary Density Operators (ADOs)

Heom.HeomAPI.getRho — Function

getRho(ados)

Return the density matrix of the reduced state (system) from a given auxiliary density operators

Parameters

ados::ADOs : the auxiliary density operators for Heom model

Returns

ρ : The density matrix of the reduced state

Heom.HeomAPI.getADO — Method

getADO(ados, idx)

Return the auxiliary density operator with a specific index from auxiliary density operators

This function equals to calling : ados[idx].

Parameters

ados::ADOs : the auxiliary density operators for Heom model

idx::Int : the index of the auxiliary density operator

Returns

ρ_idx : The auxiliary density operator

Heom.HeomAPI.getADO — Method

getADO(ados, idx_b, idx_f)

Return the auxiliary density operator with specific indices [only for mixtured (bosonic and fermionic) bath]

This function equals to calling : ados[idx_b, idx_f].

Parameters

ados::ADOs : the auxiliary density operators for Heom model

idx_b::Int : the bosonic-state index of the auxiliary density operator.

idx_f::Int : the fermionic-state index of the auxiliary density operator.

Returns

ρ_idx : The auxiliary density operator

Hierarchy Dictionary

Heom.HeomAPI.HierarchyDict — Type

struct HierarchyDict

An object which contains all dictionaries for bath-ADOs hierarchy.

The ado (or n_vector) denotes a set of integers:

\[\{ n_{11}, ..., n_{\nu k}, ... \}\]

where $n_{\nu k} \geq 0$ associated with the $k$-th exponential-expansion term in the $\nu$-th bath.

The hierarchy level ($L$) for an ado is given by $L=\sum_{\nu, k} n_{\nu k}$

Fields

idx2ado : Return the ado (n_vector) from a given index

ado2idx : Return the index from a given ado (n_vector)

lvl2idx : Return the list of indices from a given level

bathPtr : The indices for the exponential-expansion terms of the N-th bath in ado (n_vector) can be obtained by bathPtr[N]:(bathPtr[N+1] - 1)