Physical Analysis Functions

Time Evolution

Using the method based on generating Propagator

HierarchicalEOM.HeomAPI.evolution — Method

evolution(M, ρ0, Δt, steps; threshold, nonzero_tol, verbose, filename)

Solve the time evolution for auxiliary density operators based on propagator (generated by FastExpm.jl) with initial state is given in the type of density-matrix (ρ0).

This method will return the time evolution of ADOs corresponds to tlist = 0 : Δt : (Δt * steps)

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ρ0 : system initial state (density matrix)

Δt::Real : A specific time step (time interval).

steps::Int : The number of time steps

threshold::Real : Determines the threshold for the Taylor series. Defaults to 1.0e-6.

nonzero_tol::Real : Strips elements smaller than nonzero_tol at each computation step to preserve sparsity. Defaults to 1.0e-14.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

For more details, please refer to FastExpm.jl

Returns

ADOs_list : The auxiliary density operators of each time step.

HierarchicalEOM.HeomAPI.evolution — Method

evolution(M, ados, Δt, steps; threshold, nonzero_tol, verbose, filename)

Solve the time evolution for auxiliary density operators based on propagator (generated by FastExpm.jl) with initial state is given in the type of ADOs.

This method will return the time evolution of ADOs corresponds to tlist = 0 : Δt : (Δt * steps)

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ados::ADOs : initial auxiliary density operators

Δt::Real : A specific time step (time interval).

steps::Int : The number of time steps

threshold::Real : Determines the threshold for the Taylor series. Defaults to 1.0e-6.

nonzero_tol::Real : Strips elements smaller than nonzero_tol at each computation step to preserve sparsity. Defaults to 1.0e-14.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

For more details, please refer to FastExpm.jl

Returns

ADOs_list : The auxiliary density operators of each time step.

Using the method based on ODEs : OrdinaryDiffEq.jl

HierarchicalEOM.HeomAPI.evolution — Method

evolution(M, ρ0, tlist; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators based on ordinary differential equations with initial state is given in the type of density-matrix (ρ0).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ρ0 : system initial state (density matrix)

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

HierarchicalEOM.HeomAPI.evolution — Method

evolution(M, ados, tlist; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators based on ordinary differential equations with initial state is given in the type of ADOs.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ados::ADOs : initial auxiliary density operators

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

Support for time-dependent system Hamiltonian using the method based on ODEs : OrdinaryDiffEq.jl

HierarchicalEOM.HeomAPI.evolution — Function

evolution(M, ρ0, tlist, H, param; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators with time-dependent system Hamiltonian based on ordinary differential equations with initial state is given in the type of density-matrix (ρ0).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model (with time-independent system Hamiltonian)

ρ0 : system initial state (density matrix)

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

H::Function : a function for time-dependent part of system Hamiltonian. The function will be called by H(param, t) and should return the time-dependent part system Hamiltonian matrix at time t with AbstractMatrix type.

param::Tuple: the tuple of parameters which is used to call H(param, t) for the time-dependent system Hamiltonian. Default to empty tuple ().

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

HierarchicalEOM.HeomAPI.evolution — Function

evolution(M, ados, tlist, H, param; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators with time-dependent system Hamiltonian based on ordinary differential equations with initial state is given in the type of ADOs.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model (with time-independent system Hamiltonian)

ados::ADOs : initial auxiliary density operators

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

H::Function : a function for time-dependent part of system Hamiltonian. The function will be called by H(param, t) and should return the time-dependent part system Hamiltonian matrix at time t with AbstractMatrix type.

param::Tuple: the tuple of parameters which is used to call H(param, t) for the time-dependent system Hamiltonian. Default to empty tuple ().

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file "filename.jld2" during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

Steady State

Using the method based on LinearSolve.jl

HierarchicalEOM.HeomAPI.SteadyState — Method

SteadyState(M; solver, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on LinearSolve.jl (i.e., solving $x$ where $A \times x = b$).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model, where the parity should be :even.

solver : solver in package LinearSolve.jl. Default to UMFPACKFactorization().

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to LinearSolve.jl

Returns

::ADOs : The steady state of auxiliary density operators.

Using the method based on ODEs : OrdinaryDiffEq.jl

HierarchicalEOM.HeomAPI.SteadyState — Method

SteadyState(M, ρ0; solver, reltol, abstol, maxiters, save_everystep, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on time evolution (ordinary differential equations) with initial state is given in the type of density-matrix (ρ0).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model, where the parity should be :even.

ρ0 : system initial state (density matrix)

solver : The ODE solvers in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

::ADOs : The steady state of auxiliary density operators.

HierarchicalEOM.HeomAPI.SteadyState — Method

SteadyState(M, ados; solver, reltol, abstol, maxiters, save_everystep, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on time evolution (ordinary differential equations) with initial state is given in the type of ADOs.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model, where the parity should be :even.

ados::ADOs : initial auxiliary density operators

solver : The ODE solvers in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-3.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-6.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

::ADOs : The steady state of auxiliary density operators.

Spectrum

HierarchicalEOM.Spectrum.spectrum — Function

spectrum(M, ρ, op, ω_list; solver, verbose, filename, SOLVEROptions...)

Calculate spectrum for the system.

To calculate spectrum for bosonic systems (usually known as power spectrum):

\[\pi S(\omega)=\textrm{Re}\left\{\int_0^\infty dt \langle a^\dagger(t) a(0)\rangle e^{-i\omega t}\right\},\]

remember to set the parameters:

M::AbstractHEOMMatrix: should be :even parity

op: the (annihilation) operator $a$ for bosonic system as shown above

To calculate spectrum for fermionic systems (usually known as density of states):

\[ \pi A(\omega)=\textrm{Re}\left\{\int_0^\infty dt \left[\langle d(t) d^\dagger(0)\rangle^* + \langle d^\dagger(t) d(0)\rangle \right] e^{-i\omega t}\right\},\]

remember to set the parameters:

M::AbstractHEOMMatrix: should be :odd parity

op: the (annihilation) operator $d$ for fermionic system as shown above

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model.

ρ : the system density matrix or the auxiliary density operators.

op : The annihilation operator acting on the system.

ω_list::AbstractVector : the specific frequency points to solve.

solver : solver in package LinearSolve.jl. Default to UMFPACKFactorization().

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the value of spectrum for each ω will be saved into the file "filename.txt" during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to LinearSolve.jl

Returns

spec::AbstractVector : the spectrum list corresponds to the specified ω_list