Bath Module

Bosonic Bath

Heom.Bath.BosonBath — Type

struct BosonBath <: AbstractBath

An object which describes the interaction between system and bosonic bath

Fields

bath : the different boson-bath-type objects which describes the interaction between system and bosonic bath

op : The system operator according to the system-bosonic-bath interaction.

dim : the dimension of the coupling operator (should be equal to the system dimension).

Nterm : the number of exponential-expansion term of correlation functions

δ : The approximation discrepancy which is used for adding the terminator to HEOM matrix (see function: addTerminator!)

Methods

One can obtain the $k$-th exponent (exponential-expansion term) from bath::BosonBath by calling : bath[k]. Heom.jl also supports the following calls (methods) :

bath[1:k];   # returns a vector which contains the exponents from the `1`-st to the `k`-th term.
bath[1:end]; # returns a vector which contains all the exponential-expansion terms
bath[:];     # returns a vector which contains all the exponential-expansion terms
from b in bath
    # do something
end

Heom.Bath.BosonBath — Method

BosonBath(op, η, γ, δ=0.0; combine=true)

Generate BosonBath object for the case where real part and imaginary part of the correlation function are combined.

Parameters

op : The system operator according to the system-bosonic-bath interaction.

η::Vector{Ti<:Number} : the coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ::Vector{Tj<:Number} : the coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

δ::Number : The approximation discrepancy (Default to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator!)

combine::Bool : Whether to combine the exponential-expansion terms with the same frequency. Defaults to true.

Heom.Bath.BosonBath — Method

BosonBath(op, η_real, γ_real, η_imag, γ_imag, δ=0.0; combine=true)

Generate BosonBath object for the case where the correlation function splits into real part and imaginary part.

Parameters

op : The system operator according to the system-bosonic-bath interaction.

η_real::Vector{Ti<:Number} : the real part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_real::Vector{Tj<:Number} : the real part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_imag::Vector{Tk<:Number} : the imaginary part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_imag::Vector{Tl<:Number} : the imaginary part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

δ::Number : The approximation discrepancy (Default to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator!)

combine::Bool : Whether to combine the exponential-expansion terms with the same frequency. Defaults to true.

Heom.Bath.bosonReal — Type

struct bosonReal <: AbstractBosonBath

A bosonic bath for the real part of bath correlation function

Fields

Comm : the super-operator (commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the real part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ : the real part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Nterm : the number of exponential-expansion term of correlation function

Heom.Bath.bosonReal — Method

bosonReal(op, η_real, γ_real)

Generate bosonic bath for the real part of bath correlation function

Parameters

op : The system operator according to the system-bosonic-bath interaction.

η_real::Vector{Ti<:Number} : the real part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_real::Vector{Tj<:Number} : the real part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Heom.Bath.bosonImag — Type

struct bosonImag <: AbstractBosonBath

A bosonic bath for the imaginary part of bath correlation function

Fields

Comm : the super-operator (commutator) for the coupling operator.

anComm : the super-operator (anti-commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the imaginary part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ : the imaginary part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Nterm : the number of exponential-expansion term of correlation function

Heom.Bath.bosonImag — Method

bosonImag(op, η_imag, γ_imag)

Generate bosonic bath for the imaginary part of correlation function

Parameters

op : The system operator according to the system-bosonic-bath interaction.

η_imag::Vector{Ti<:Number} : the imaginary part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_imag::Vector{Tj<:Number} : the imaginary part of coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Heom.Bath.bosonRealImag — Type

sturct bosonRealImag <: AbstractBosonBath

A bosonic bath which the real part and imaginary part of the bath correlation function are combined

Fields

Comm : the super-operator (commutator) for the coupling operator.

anComm : the super-operator (anti-commutator) for the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η_real : the real part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_imag : the imaginary part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ : the coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Nterm : the number of exponential-expansion term of correlation function

Heom.Bath.bosonRealImag — Method

bosonRealImag(op, η_real, η_imag, γ)

Generate bosonic bath which the real part and imaginary part of the bath correlation function are combined

Parameters

op : The system operator according to the system-bosonic-bath interaction.

η_real::Vector{Ti<:Number} : the real part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_imag::Vector{Tj<:Number} : the imaginary part of coefficients $\eta_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ::Vector{Tk<:Number} : the coefficients $\gamma_i$ in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Fermionic Bath

Heom.Bath.FermionBath — Type

struct FermionBath <: AbstractBath

An object which describes the interaction between system and fermionic bath

Fields

bath : the different fermion-bath-type objects which describes the interaction

op : The system "emission" operator according to the system-fermionic-bath interaction.

dim : the dimension of the coupling operator (should be equal to the system dimension).

Nterm : the number of exponential-expansion term of correlation functions

δ : The approximation discrepancy which is used for adding the terminator to HEOM matrix (see function: addTerminator!)

Methods

One can obtain the $k$-th exponent (exponential-expansion term) from bath::FermionBath by calling : bath[k]. Heom.jl also supports the following calls (methods) :

bath[1:k];   # returns a vector which contains the exponents from the `1`-st to the `k`-th term.
bath[1:end]; # returns a vector which contains all the exponential-expansion terms
bath[:];     # returns a vector which contains all the exponential-expansion terms
from b in bath
    # do something
end

Heom.Bath.FermionBath — Method

FermionBath(op, η_absorb, γ_absorb, η_emit, γ_emit, δ=0.0)

Generate FermionBath object

Parameters

op : The system "emission" operator according to the system-fermionic-bath interaction.

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_absorb::Vector{Tj<:Number} : the coefficients $\gamma_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_emit::Vector{Tk<:Number} : the coefficients $\eta_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_emit::Vector{Tl<:Number} : the coefficients $\gamma_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

δ::Number : The approximation discrepancy (Defaults to 0.0) which is used for adding the terminator to HEOM matrix (see function: addTerminator!)

Heom.Bath.fermionAbsorb — Type

struct fermionAbsorb <: AbstractFermionBath

An object which describes the absorption of the system in the interaction

Fields

spre : the super-operator (right side operator multiplication) for the coupling operator.

spost : the super-operator (left side operator multiplication) for the coupling operator.

spreD : the super-operator (right side operator multiplication) for the adjoint of the coupling operator.

spostD : the super-operator (left side operator multiplication) for the adjoint of the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ : the coefficients $\gamma_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_emit : the coefficients $\eta_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Nterm : the number of exponential-expansion term of correlation function

Heom.Bath.fermionAbsorb — Method

fermionAbsorb(op, η_absorb, γ_absorb, η_emit)

Generate fermionic bath which describes the absorption of the system in the interaction

Parameters

op : The system absorption operator according to the system-fermionic-bath interaction.

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_absorb::Vector{Tj<:Number} : the coefficients $\gamma_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_emit::Vector{Tk<:Number} : the coefficients $\eta_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Heom.Bath.fermionEmit — Type

struct fermionEmit <: AbstractFermionBath

An object which describes the emission of the system in the interaction

Fields

spre : the super-operator (right side operator multiplication) for the coupling operator.

spost : the super-operator (left side operator multiplication) for the coupling operator.

spreD : the super-operator (right side operator multiplication) for the adjoint of the coupling operator.

spostD : the super-operator (left side operator multiplication) for the adjoint of the coupling operator.

dim : the dimension of the coupling operator (should be equal to the system dimension).

η : the coefficients $\eta_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ : the coefficients $\gamma_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_absorb : the coefficients $\eta_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Nterm : the number of exponential-expansion term of correlation function

Heom.Bath.fermionEmit — Method

fermionEmit(op, η_emit, γ_emit, η_absorb)

Generate fermionic bath which describes the absorption of the system in the interaction

Parameters

op : The system emission operator according to the system-fermionic-bath interaction.

η_emit::Vector{Ti<:Number} : the coefficients $\eta_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

γ_emit::Vector{Ti<:Number} : the coefficients $\gamma_i$ for emission in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

η_absorb::Vector{Ti<:Number} : the coefficients $\eta_i$ for absorption in bath correlation functions ($\sum_i \eta_i e^{-\gamma_i t}$).

Exponent

Heom.Bath.Exponent — Type

struct Exponent

An object which describes a single exponential-expansion term (naively, an excitation mode) within the decomposition of the bath correlation functions.

The expansion of a bath correlation function can be expressed as : $\sum_i \eta_i e^{-\gamma_i t}$.

Fields

op : The coupling operator according to system-bath interaction.

η::Number : the coefficient $\eta_i$ in bath correlation functions.

γ::Number : the coefficient $\gamma_i$ in bath correlation functions.

types::String : The type-tag of the exponent.

The types different types of the Exponent:

"bR" : from real part of bosonic correlation function

"bI" : from imaginary part of bosonic correlation function

"bRI" : from combined (real and imaginary part) bosonic bath correlation function

"fA" : from the fermionic bath which describes the absorption

"fE" : from the fermionic bath which describes the emission