Heom API Module

Heom liouvillian superoperator matrices

Boson Heom matrix

Heom.HeomAPI.M_Boson — Type

struct M_Boson <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for bosonic bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

tier : the tier (cutoff level) for the hierarchy

dim : the dimension of system

N : the number of total ADOs

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix (restrict to :none for boson)

bath::Vector{BosonBath} : the vector which stores all BosonBath objects

hierarchy::HierarchyDict: the object which contains all dictionaries for boson-bath-ADOs hierarchy.

Heom.HeomAPI.M_Boson — Method

M_Boson(Hsys, tier, Bath; threshold=0.0, verbose=true)

Generate the boson-type Heom liouvillian superoperator matrix

Parameters

Hsys : The time-independent system Hamiltonian

tier::Int : the tier (cutoff) for the bath

Bath::Vector{BosonBath} : objects for different bosonic baths

threshold::Real : The threshold of the importance value (see Ref. [1]). Defaults to 0.0.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

[1] Phys. Rev. B 88, 235426 (2013)

Fermion Heom matrix

Heom.HeomAPI.M_Fermion — Type

struct M_Fermion <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for fermionic bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

tier : the tier (cutoff level) for the hierarchy

dim : the dimension of system

N : the number of total ADOs

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix

bath::Vector{FermionBath} : the vector which stores all FermionBath objects

hierarchy::HierarchyDict: the object which contains all dictionaries for fermion-bath-ADOs hierarchy.

Heom.HeomAPI.M_Fermion — Type

M_Fermion(Hsys, tier, Bath, parity=:even; threshold=0.0, verbose=true)

Generate the fermion-type Heom liouvillian superoperator matrix

Parameters

Hsys : The time-independent system Hamiltonian

tier::Int : the tier (cutoff) for the bath

Bath::Vector{FermionBath} : objects for different fermionic baths

parity::Symbol : The parity symbol of the density matrix (either :odd or :even). Defaults to :even.

threshold::Real : The threshold of the importance value (see Ref. [1]). Defaults to 0.0.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

[1] Phys. Rev. B 88, 235426 (2013)

Boson-Fermion Heom matrix

Heom.HeomAPI.M_Boson_Fermion — Type

struct M_Boson_Fermion <: AbstractHEOMMatrix

Heom liouvillian superoperator matrix for mixtured (bosonic and fermionic) bath

Fields

data : the sparse matrix of HEOM liouvillian superoperator

Btier : the tier (cutoff level) for bosonic hierarchy

Ftier : the tier (cutoff level) for fermionic hierarchy

dim : the dimension of system

N : the number of total ADOs

sup_dim : the dimension of system superoperator

parity : the parity of the density matrix

Bbath::Vector{BosonBath} : the vector which stores all BosonBath objects

Fbath::Vector{FermionBath} : the vector which stores all FermionBath objects

hierarchy::MixHierarchyDict: the object which contains all dictionaries for mixed-bath-ADOs hierarchy.

Heom.HeomAPI.M_Boson_Fermion — Type

M_Boson_Fermion(Hsys, Btier, Ftier, Bbath, Fbath, parity=:even; threshold=0.0, verbose=true)

Generate the boson-fermion-type Heom liouvillian superoperator matrix

Parameters

Hsys : The time-independent system Hamiltonian

Btier::Int : the tier (cutoff) for the bosonic bath

Ftier::Int : the tier (cutoff) for the fermionic bath

Bbath::Vector{BosonBath} : objects for different bosonic baths

Fbath::Vector{FermionBath} : objects for different fermionic baths

parity::Symbol : The parity symbol of the density matrix (either :odd or :even). Defaults to :even.

threshold::Real : The threshold of the importance value (see Ref. [1, 2]). Defaults to 0.0.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

[1] Phys. Rev. B 88, 235426 (2013) [2] Phys. Rev. B 103, 235413 (2021)

Functions for Heom Matrices

Base.size — Method

size(M::AbstractHEOMMatrix)

Returns the size of the Heom liouvillian superoperator matrix

Heom.HeomAPI.Propagator — Function

Propagator(M, Δt; threshold, nonzero_tol)

Use FastExpm.jl to calculate the propagator matrix from a given Heom liouvillian superoperator matrix $M$ with a specific time step $Δt$. That is, $\exp(M * \Delta t)$.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

Δt::Real : A specific time step (time interval).

threshold::Real : Determines the threshold for the Taylor series. Defaults to 1.0e-6.

nonzero_tol::Real : Strips elements smaller than nonzero_tol at each computation step to preserve sparsity. Defaults to 1.0e-14.

For more details, please refer to FastExpm.jl

Returns

::SparseMatrixCSC{ComplexF64, Int64} : the propagator matrix

Heom.HeomAPI.addDissipator — Function

addDissipator(M, jumpOP)

Adding dissipator to a given HEOM matrix.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

jumpOP::AbstractVector : The collapse (jump) operators to add. Defaults to empty vector [].

Return

M_new::AbstractHEOMMatrix : the new HEOM liouvillian superoperator matrix

Heom.HeomAPI.addTerminator — Function

addTerminator(M, Bath)

Adding terminator to a given HEOM matrix.

The terminator is a liouvillian term representing the contribution to the system-bath dynamics of all exponential-expansion terms beyond Bath.Nterm

The difference between the true correlation function and the sum of the Bath.Nterm-exponential terms is approximately 2 * δ * dirac(t). Here, δ is the approximation discrepancy and dirac(t) denotes the Dirac-delta function.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

Bath::Union{BosonBath, FermionBath} : The bath object which contains the approximation discrepancy δ

Return

M_new::AbstractHEOMMatrix : the new HEOM liouvillian superoperator matrix

Auxiliary Density Operators (ADOs)

Heom.HeomAPI.ADOs — Type

mutable struct ADOs

The Auxiliary Density Operators for Heom model.

Fields

data : the vectorized auxiliary density operators

dim : the dimension of the system

N : the number of auxiliary density operators

Methods

One can obtain the density matrix for specific index (idx) by calling : ados[idx]. Heom.jl also supports the following calls (methods) :

length(ados);  # returns the total number of `ADOs`
ados[1:idx];   # returns a vector which contains the `ADO` (in matrix form) from index `1` to `idx`
ados[1:end];   # returns a vector which contains all the `ADO` (in matrix form)
ados[:];       # returns a vector which contains all the `ADO` (in matrix form)
from rho in ados  # iteration
    # do something
end

Heom.HeomAPI.ADOs — Method

ADOs(V, N)

Gernerate the object of auxiliary density operators for Heom model.

Parameters

V::AbstractVector : the vectorized auxiliary density operators

N::Int : the number of auxiliary density operators.

Functions for Auxiliary Density Operators

Base.length — Method

length(A::ADOs)

Returns the total number of the Auxiliary Density Operators (ADOs)

Heom.HeomAPI.getRho — Function

getRho(ados)

Return the density matrix of the reduced state (system) from a given auxiliary density operators

Parameters

ados::ADOs : the auxiliary density operators for Heom model

Returns

ρ : The density matrix of the reduced state

Heom.HeomAPI.getADO — Method

getADO(ados, idx)

Return the auxiliary density operator with a specific index from auxiliary density operators

This function equals to calling : ados[idx].

Parameters

ados::ADOs : the auxiliary density operators for Heom model

idx::Int : the index of the auxiliary density operator

Returns

ρ_idx : The auxiliary density operator

Hierarchy Dictionary

Heom.HeomAPI.Nvec — Type

struct Nvec

An object which describes the repetition number of each multi-index ensembles in auxiliary density operators.

The n_vector ($\vec{n}$) denotes a set of integers:

\[\{ n_{1,1}, ..., n_{\alpha, k}, ... \}\]

associated with the $k$-th exponential-expansion term in the $\alpha$-th bath. If $n_{\alpha, k} = 3$ means that the multi-index ensemble $\{\alpha, k\}$ appears three times in the multi-index vector of ADOs (see the notations in our paper).

The hierarchy level ($L$) for an n_vector is given by $L=\sum_{\alpha, k} n_{\alpha, k}$

Fields

data : the n_vector

level : The level L for the n_vector

Methods

One can obtain the repetition number for specific index (idx) by calling : n_vector[idx]. To obtain the corresponding tuple $(\alpha, k)$ for a given index idx, see bathPtr in HierarchyDict for more details.

Heom.jl also supports the following calls (methods) : ``julia length(n_vector); # returns the length ofNvecn_vector[1:idx]; # returns a vector which contains the excitation number ofnvectorfrom index1toidx` nvector[1:end]; # returns a vector which contains all the excitation number of n_vector nvector[:]; # returns a vector which contains all the excitation number of `nvector` from n in n_vector # iteration # do something end

Heom.HeomAPI.HierarchyDict — Type

struct HierarchyDict <: AbstractHierarchyDict

An object which contains all dictionaries for pure (bosonic or fermionic) bath-ADOs hierarchy.

Fields

idx2nvec : Return the Nvec from a given index

nvec2idx : Return the index from a given Nvec

lvl2idx : Return the list of indices from a given level

bathPtr : Records the tuple $(\alpha, k)$ for each position in Nvec, where $\alpha$ and $k$ represents the $k$-th exponential-expansion term of the $\alpha$-th bath.

Heom.HeomAPI.MixHierarchyDict — Type

struct MixHierarchyDict <: AbstractHierarchyDict

An object which contains all dictionaries for mixed (bosonic and fermionic) bath-ADOs hierarchy.

Fields

idx2nvec : Return the tuple (Nvec_b, Nvec_f) from a given index, where b represents boson and f represents fermion

nvec2idx : Return the index from a given tuple (Nvec_b, Nvec_f), where b represents boson and f represents fermion

Blvl2idx : Return the list of indices from a given bosonic level (excitation)

Flvl2idx : Return the list of indices from a given fermionic level (excitation)

bosonPtr : Records the tuple $(\alpha, k)$ for each position in Nvec_b, where $\alpha$ and $k$ represents the $k$-th exponential-expansion term of the $\alpha$-th bosonic bath.

fermionPtr : Records the tuple $(\alpha, k)$ for each position in Nvec_f, where $\alpha$ and $k$ represents the $k$-th exponential-expansion term of the $\alpha$-th fermionic bath.

Heom.HeomAPI.getIndexEnsemble — Function

getIndexEnsemble(nvec, bathPtr)

Search for all the multi-index ensemble $(\alpha, k)$ where $\alpha$ and $k$ represents the $k$-th exponential-expansion term in the $\alpha$-th bath.

Parameters

nvec::Nvec : An object which records the repetition number of each multi-index ensembles in ADOs.

bathPtr::Vector{Tuple{Int, Int}}: This can be obtained from HierarchyDict.bathPtr, MixHierarchyDict.bosonPtr, or MixHierarchyDict.fermionPtr.

Returns

Vector{Tuple{Int, Int, Int}}: a vector (list) of the tuples $(\alpha, k, n)$.

Example

Here is an example to use Bath, Exponent, HierarchyDict, and getIndexEnsemble together:

L::M_Fermion;          # suppose this is a fermion type of HEOM liouvillian superoperator matrix you create
HDict = L.hierarchy;   # the hierarchy dictionary
ados = SteadyState(L); # the stationary state (ADOs) for L 

# Let's consider all the ADOs for first level
idx_list = HDict.lvl2idx[1];

for idx in idx_list
    ρ1 = ados[idx]  # one of the 1st-level ADO
    nvec = HDict.idx2nvec[idx]  # the nvec corresponding to ρ1
    
    for (α, k, n) in getEnsemble(nvec, HDict.bathPtr)
        α  # index of the bath
        k  # the index of the exponential-expansion term in α-th bath
        n  # the repetition number of the ensemble \{\alpha, k\} in ADOs
        exponent = L.bath[α][k]  # the k-th exponential-expansion term in α-th bath

        # do some calculations you want
    end
end