Physical Analysis Functions

Time Evolution

Using the method based on generating Propagator

Heom.HeomAPI.evolution — Method

evolution(M, ρ0, Δt, steps; threshold, nonzero_tol, verbose, filename)

Solve the time evolution for auxiliary density operators based on propagator (generated by FastExpm.jl) with initial state is given in the type of density-matrix (ρ0).

This method will return the time evolution of ADOs corresponds to tlist = 0 : Δt : (Δt * steps)

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ρ0 : system initial state (density matrix)

Δt::Real : A specific time step (time interval).

steps::Int : The number of time steps

threshold::Real : Determines the threshold for the Taylor series. Defaults to 1.0e-6.

nonzero_tol::Real : Strips elements smaller than nonzero_tol at each computation step to preserve sparsity. Defaults to 1.0e-14.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file during the solving process.

For more details, please refer to FastExpm.jl

Returns

ADOs_list : The auxiliary density operators of each time step.

Heom.HeomAPI.evolution — Method

evolution(M, ados, Δt, steps; threshold, nonzero_tol, verbose, filename)

Solve the time evolution for auxiliary density operators based on propagator (generated by FastExpm.jl) with initial state is given in the type of ADOs.

This method will return the time evolution of ADOs corresponds to tlist = 0 : Δt : (Δt * steps)

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ados::ADOs : initial auxiliary density operators

Δt::Real : A specific time step (time interval).

steps::Int : The number of time steps

threshold::Real : Determines the threshold for the Taylor series. Defaults to 1.0e-6.

nonzero_tol::Real : Strips elements smaller than nonzero_tol at each computation step to preserve sparsity. Defaults to 1.0e-14.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file during the solving process.

For more details, please refer to FastExpm.jl

Returns

ADOs_list : The auxiliary density operators of each time step.

Using the method based on ODEs : OrdinaryDiffEq.jl

Heom.HeomAPI.evolution — Method

evolution(M, ρ0, tlist; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators based on ordinary differential equations with initial state is given in the type of density-matrix (ρ0).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ρ0 : system initial state (density matrix)

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

Heom.HeomAPI.evolution — Method

evolution(M, ados, tlist; solver, reltol, abstol, maxiters, save_everystep, verbose, filename, SOLVEROptions...)

Solve the time evolution for auxiliary density operators based on ordinary differential equations with initial state is given in the type of ADOs.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ados::ADOs : initial auxiliary density operators

tlist::AbstractVector : Denote the specific time points to save the solution at, during the solving process.

solver : solver in package DifferentialEquations.jl. Default to DP5().

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the ADOs at each time point will be saved into the JLD2 file during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

ADOs_list : The auxiliary density operators in each time point.

Steady State

Using the method based on LinearSolve.jl

Heom.HeomAPI.SteadyState — Method

SteadyState(M; solver, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on LinearSolve.jl (i.e., solving $x$ where $A × x = b$).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model, where the parity should be either :none (boson) or :even (fermion).

solver : solver in package LinearSolve.jl. Default to UMFPACKFactorization().

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to LinearSolve.jl

Returns

::ADOs : The steady state of auxiliary density operators.

Using the method based on ODEs : OrdinaryDiffEq.jl

Heom.HeomAPI.SteadyState — Method

SteadyState(M, ρ0; solver, reltol, abstol, maxiters, save_everystep, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on time evolution (ordinary differential equations) with initial state is given in the type of density-matrix (ρ0).

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ρ0 : system initial state (density matrix)

solver : The ODE solvers in package DifferentialEquations.jl. Default to FBDF(autodiff=false).

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

::ADOs : The steady state of auxiliary density operators.

Heom.HeomAPI.SteadyState — Method

SteadyState(M, ados; solver, reltol, abstol, maxiters, save_everystep, verbose, SOLVEROptions...)

Solve the steady state of the auxiliary density operators based on time evolution (ordinary differential equations) with initial state is given in the type of ADOs.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model

ados::ADOs : initial auxiliary density operators

solver : The ODE solvers in package DifferentialEquations.jl. Default to FBDF(autodiff=false).

reltol::Real : Relative tolerance in adaptive timestepping. Default to 1.0e-6.

abstol::Real : Absolute tolerance in adaptive timestepping. Default to 1.0e-8.

maxiters::Real : Maximum number of iterations before stopping. Default to 1e5.

save_everystep::Bool : Saves the result at every step. Defaults to false.

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to DifferentialEquations.jl

Returns

::ADOs : The steady state of auxiliary density operators.

Spectrum

Power Spectral Density (for bosonic type Bath)

Heom.Spectrum.PSD — Function

PSD(M, ρ, op, ω_list; solver, verbose, filename, SOLVEROptions...)

Calculate power spectral density.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model (the parity must be :none or :even.)

ρ : the system density matrix or the auxiliary density operators.

op : The system operator for the two-time correlation function in frequency domain.

ω_list::AbstractVector : the specific frequency points to solve.

solver : solver in package LinearSolve.jl. Default to UMFPACKFactorization().

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the value of psd for each ω will be saved into the file during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to LinearSolve.jl

Returns

psd::AbstractVector : power spectral density

Density of States (for fermionic type Bath)

Heom.Spectrum.DOS — Function

DOS(M, ρ, op, ω_list; solver, verbose, filename, SOLVEROptions...)

Calculate density of states.

Parameters

M::AbstractHEOMMatrix : the matrix given from HEOM model (the parity must be :odd.)

ρ : the system density matrix or the auxiliary density operators.

op : The system operator for the two-time correlation function in frequency domain.

ω_list::AbstractVector : the specific frequency points to solve.

solver : solver in package LinearSolve.jl. Default to UMFPACKFactorization().

verbose::Bool : To display verbose output and progress bar during the process or not. Defaults to true.

filename::String : If filename was specified, the value of dos for each ω will be saved into the file during the solving process.

SOLVEROptions : extra options for solver

For more details about solvers and extra options, please refer to LinearSolve.jl

Returns

dos::AbstractVector : density of state